+++ packages/nuget-exe/.gitignore
@@ -1,0 +1,1 @@
+/nuget-*.exe
