+++ packages/nuget-exe/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "nuget-exe",
+  "version": "4.9.1",
+  "description": "Repackaged NuGet.exe bianry",
+  "main": "nuget-4.9.2.exe",
+  "repository": "https://github.com/microsoft/react-native-windows",
+  "license": "MIT",
+  "scripts": {
+    "build": "just-scripts build",
+    "clean": "just-scripts clean"
+  },
+  "devDependencies": {
+    "just-scripts": "^0.36.1"
+  },
+  "files": [
+    "nuget-4.9.2.exe"
+  ]
+}
