+++ vnext/package.json
@@ -32,6 +32,7 @@
     "mustache": "^4.0.1",
     "ora": "^3.4.0",
     "prop-types": "^15.7.2",
+    "nuget-exe": "4.9.1",
     "regenerator-runtime": "^0.13.2",
     "semver": "^7.1.3",
     "shelljs": "^0.7.8",
